refactor(binding-kafka): rename DescribeCluster api classes with Kafka prefix
api package class names begin with Kafka* elsewhere in the package
(KafkaCreateTopicsRequest, KafkaDeleteTopicsRequest); the DescribeCluster
response types added in this PR were the outlier. Renames scoped only to
classes this PR introduced -- CreateTopicsResponse/DeleteTopicsResponse
and their siblings predate this PR and are left as-is.
